fix: alterações de tipos sob parâmetros da entity Product, restart no banco de dados

+++ src/main/java/com/fcolucasvieira/auth/dto/product/ProductRequestDTO.java
@@ -3,9 +3,11 @@ package com.fcolucasvieira.auth.dto.product;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
+
 public record ProductRequestDTO(
         @NotBlank
         String name,
         @NotNull
-        Integer price
+        BigDecimal price
 ) {}
